refactor: Refatorado o envio de e-mail utilizando template.
